Testing a potential error in Ch11
This is to do with the additional visit to the cash machine recently added to the end of Ch11.
Potentially jumps from Ch11 to Ch12 after cash withdrawal.
Likely due to the 'buyCash' component and the 'current module' or something around that.
Thought I had figured it out just using an html checkfile.